@@ -3,9 +3,13 @@
   "version": "1.0.0",
   "description": "experimentations with react and related technologies",
   "dependencies": {
+    "babel-runtime": "^5.8.25",
     "immutable": "^3.7.5",
     "marked": "^0.3.5",
-    "react": "^0.13.3"
+    "react": "^0.14.0-rc1",
+    "react-dom": "^0.14.0-rc1",
+    "rx": "^4.0.0",
+    "rx-react": "^0.3.0-beta2"
   },
   "devDependencies": {
     "babel-jest": "^5.3.0",
