@@ -1,38 +1,32 @@
 /* @flow */
-import React from 'react/addons';
-import Immutable from 'immutable';
-import commentData from './commentData';
+import React from 'react';
+import RxReact from 'rx-react';
 import CommentForm from './commentForm';
 import CommentList from './commentList';
-import { CommentRecord, toListOfComments } from './commentRecord';
+import { commentListStream, makeComment } from './commentRecord';
 
-export default React.createClass({
-  mixins: [React.addons.PureRenderMixin],
-  getInitialState: function () {
-    return { data: toListOfComments([]) }
-  },
-  handleCommentSubmit: function (comment: Immutable.Record) {
-    const comments: any/*Immutable.List*/ = this.state.data;
-    const newComment = new CommentRecord(comment);
-    this.replaceState({
-      data: comments.concat([ newComment.set('id', `local_${Date.now()}`) ])
+export class CommentComponent extends RxReact.Component {
+  getStateStream () {
+    return commentListStream.map(list => {
+      return { data: list }
     });
-  },
-  loadComments: function () {
-    this.replaceState({
-      data: toListOfComments(commentData)
-    });
-  },
-  componentDidMount: function () {
-    this.loadComments();
-  },
-  render: function (): ReactElement {
+  }
+
+  componentWillMount () {
+    super.componentWillMount();
+    this.makeComment = RxReact.FuncSubject.create();
+    this.makeComment
+      .forEach(makeComment);
+  }
+
+  render (): ReactElement {
     return (
       <div className="commentBox">
         <h1>Comments</h1>
-        <CommentList data={this.state.data}/>
-        <CommentForm onCommentSubmit={this.handleCommentSubmit} />
+        <CommentList data={this.state.data} />
+        <CommentForm onCommentSubmit={this.makeComment} />
       </div>
     );
   }
-});
+}
+
